requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.8.0
django-phonenumber-field==8.0.0
phonenumbers==8.13.0
cryptography==42.0.0
python-jose==1.8.2
python-multipart==0.0.6
Pillow==11.3.0
celery==5.4.0
redis==5.0.0
drf-yasg==1.21.8
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.3
email-validator==2.3.0
